pyproject.toml
[project]
name = "churn-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "xgboost",
    "optuna",
    "pandas",
    "scikit-learn",
    "joblib",
    "numpy",
    "pydantic>=2.0",
    "pyyaml",
    "google-cloud-bigquery",
    "db-dtypes",
    "kfp",
    "google-cloud-aiplatform",
    "pre-commit",
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.terraform
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = [".venv/*", "__pycache__/*", ".terraform/*"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".venv", "__pycache__", ".terraform"]
